// app-old/src/types/app-translation.ts
export interface AppTranslation {
  dashboard: string;
  settings: string;
  information: string;

  save: string;
  apiKey: string;
  apiKeyDescription: string;
  apiKeySaved: string;
  baseUrl: string;
  baseUrlDescription: string;
  baseUrlSaved: string;
  name: string;
  nameDescription: string;
  nameSaved: string;
  emails: string;
  emailsDescription: string;
  emailsSaved: string;
  language: string;
  email: string;
  emptyList: string;
  unsavedChanges: string;
  nextAlert: string;
  alertCountdown: string;
  timerElapsed: string;
  cancel: string;
  timerUpdated: string;
  allow: string;
  notifications1: string;
  notifications2: string;

  connectionError: string;
  permissionRequired: string;
  unknownError: string;
  requestError: string;
}